Move displayedRatingFromPreciseRating(_:totalStars:fillMode:) function from SBRatingHandler.swift to SBTouchHandler.swift and delete SBRatingHandler.swift file
